init-top/framebuffer: Load intel-agp for intelfb
After further experimentation, I discovered additional problems that my
first patch did not address, namely that

1) Some FB drivers need the AGP subsystem up and running before they
   are loaded and
2) intelfb needs intel-agp.ko, but does not have a dependency on it.

intelfb does not actually *work* on my testsystem after this (it
crashes), but unlike with plain initramfs-tools, it loads (and prints
something useful if loaded with the probeonly option). I'll try to find
out why it fails to work tomorrow; it's probably an unrelated issue.

[ make the patch applyable, probably whitespace damaged, fix comments,
  no need to pass -q to modprobe that is set globaly -maks ]

(closes: #416063, #455876)

Signed-off-by: maximilian attems <maks@debian.org>