update-initramfs: fix ro-mounted /boot check (closes: #458772)
Hi Kernel-Team,

i stumbled over this bug when I wanted to update my initrd while I had
some isos loop-mounted like this:

$ cat /proc/mounts |grep boot
/dev/hda1 /boot ext2 rw 0 0
/dev/loop57 /srv/pub/iso_loop/sys/bootrescue-cd-20071129-59 iso9660 ro 0
0
/dev/loop58 /srv/pub/iso_loop/sys/bootrescue-cd-dev iso9660 ro 0 0
/dev/loop59 /srv/pub/iso_loop/sys/bootrescue-dvd-20071129-59 iso9660 ro
0 0
/dev/loop60 /srv/pub/iso_loop/sys/bootrescue-dvd-dev iso9660 ro 0 0

The check for a ro-mounted /boot partition was falsly triggered by the
mount targets of the iso-images. As you can see in the patch attached,
for the conditional awk output to work as desired, another pair of hyphen
enclosing the awk match calls is required.

HTH,

Daniel Reichelt