scripts/functions: fix configure_networking() for multiple interfaces.
IPOPTS can be assumed to get passed :::::: and thus configure
any possible interface. we have thus also to source the relevant
ipconfig output :)

based on a patch by Michal Sojka <sojkam1@fel.cvut.cz>:
"If I want several computers to boot from the same ramdisk (with NFS
root) and some computers have multiple network interfaces, the DEVICE
variable can't be set to a specific value (e.g. eth0). Ipconfig from
klibc supports a mode, where DHCP request is sent to all interfaces
and the one receiving the first DHCP offer is used. This perfectly
suits my needs, but it can't be used with initramfs scripts because
the interface name in not known in advance and the DEVICE variable is
set to something like ::::::."