update-initramfs: Initramfs generation reliability fixes.
Attached is a patch to improve the way update-initramfs handles the
generating of new or updated initramfs images for kernels. It puts in
place a few measures to ensure that if there is not enough disk space on
/boot, a previous initramfs is still in tact, to allow the booting of
that kernel. The patch applies against current git head. See the Ubuntu
specification found at the following URL for more information:
https://wiki.ubuntu.com/HardyInitramfsErrorHandling

This patch does the following:
* When generating a new initramfs, instead of copying the old initramfs to a
  backup file, it hard links it instead, so no more space is used than
  necessary. it only copies to the backup file in the event that the filesystem
  on /boot doesn't support hard links, eg FAT32 partitions.

* A new file is used to generate the initramfs. If the generation succeeds, it
  is moved to the original initramfs's location, ready to use. If it fails, the
  original initramfs is not affected, allowing it to be used at boot.

[ wrap on long line -maks ]