initramfs-tools: Fix UUID rootfs detection with 'MODULES=dep'
 When MODULES is set to 'dep' in /etc/initramfs-tools/initramfs.conf,
 update-initramfs fails to detect correctly the root filesystem(device
 and fs used). When called, this command fails with:
----------
mkinitramfs: missing rootfs root rootfs /sys entry
mkinitramfs: workaround is MODULES=most
mkinitramfs: Error please report the bug
update-initramfs: failed for /boot/initrd.img-2.6.25-2-amd64
----------

So we have 3 cases here. eval "$(mount | awk '/"... returns
	* an UUID when the rootfs is mounted with the 'LABEL=' or 'UUID='.
	 The UUID is a relative symbolic link located in /dev/disk/by-uuid/
	 pointing to the real device.

        * the real disk device.

        * /dev/root

When "root=/dev/root", we have to figure out what it is, which is
done in the "if" block. Also changing the matching regex in front
 to catch that type.

Once we got the symlink to the root device, we can use 'readlink' to
get the real device. Using 'readlink' on a file which is not a
symlink doesn't matters(case 2).