make log_begin_msg not emit trailing newline
Please make log_begin_msg not emit a trailing newline - this makes the non-
"quiet" output cleaner and results in half as many lines being emitted. For
example:

  Begin: Finding root filesytem ...
  Done.

Becomes:

  Begin: Finding root filesytem ... done.

Patch attached - it also adds a space and alters the case of "done" for
symmetry. This would be especially useful in Debian Live where we show a
large number of these messages by default.

(closes: #494257)