Fix parse_numeric() to ignore non hex root string prefixes
On OLPC machines, root is a nand device that is mounted as mtd0 (it is
neither a block device nor a char device).  The arguments passed to the
kernel are "ro root=mtd0 rootfstype=jffs2".

Unfortunately, attempting to use an initrd based upon initramfs-tools
on such a machine results in a kernel panic and a syntax error.


Begin: Mounting root file system ... /init: line 172: syntax error:
0xmtd0

The probably appears to be in parse_numeric(); the init scripts
assume that normal devices are always prefixed with /, and root=
strings that aren't are raw device numbers (prefixing them with
0x).  I'm not sure if there are other devices similar to mtd that
don't begin with a /.

How about something like the following patch? It's not foolproof,
but it at least ignores things that can't possibly be hex strings.
fixes partially #497133