allow root=mtd0 to be used; skip root checks if ROOT doesn't start with /dev.
The attached patch is needed for booting with root=mtd0.
Basically, the init scripts assume the root device should have a special
device file somewhere, and if not will wait for it to appear.  This patch
changes it so that it only waits for a device file to appear if $ROOT starts
with "/dev".  In the case of things like ROOT=LABEL=foo, the init scripts
will translate that to a device.  With ROOT=mtd0, the init scripts will not
check for a device, and will successfully mount the root filesystem.

Signed-off-by: Andres Salomon <dilinger@debian.org>