MODULES=dep fix for Xen virtual block devices
Ferenc seems to be using the xvda1=img1,xvda2=img2 scheme of Xen disks
(this is what xentools gets you) rather than the whole disk with
partition scheme.

In the xvda1/xvda2 scheme /sys/block contains:
        # ls /sys/block/
        ram0  ram1  ram10  ram11  ram12  ram13 ram14  ram15  ram2  ram3
        ram4 ram5  ram6  ram7  ram8 ram9  xvda1  xvda2

In the whole disk scheme it contains xvda as you would expect:
        # ls /sys/block/
        loop0  loop1  loop2  loop3  loop4  loop5  loop6  loop7 ram0
        ram1  ram10  ram11  ram12  ram13 ram14  ram15  ram2  ram3  ram4
        ram5  ram6  ram7  ram8 ram9  xvda

I've attached a patch which takes care of this difference.

[ remove largest suffix pattern, so that we don't stumble on big minors -maks ]

Signed-off-by: maximilian attems <maks@debian.org>