configure_networking: Raise ipconfig timeout to 180 seconds.
in 0.92m, the timeout for ipconfig in scripts/functions was set to 60
seconds. This broke our iscsi-rootfs setups.

It took me a while to debug this, but it seems that 60 seconds is not
enough for switches which have spanning-tree enabled, they need some
more time before they take up a port, in our case between 62 and 65
seconds.

(closes: #511085)