Fix boot with LABEL containting one or several '/'
As there is no /dev/disk/by-label//, we need to follow udev
on  escape the '/' character with  \x2f

don't rely on sed, as we might not have it in initramfs.

(closes: #489008)

Reported-by: Josh Triplett <josh@freedesktop.org>
Tested-by: Andres Salomon <dilinger@debian.org>

thanks for dilinger to pushing the first easy solution to a complete,
so we don't have to revisit the case.