init: export ip root param for configure_networking()
while at it cleanup the variable name that came due to legacy
naming nfsopts in a very early initramfs-tools release.

"The local-top/iscsi initramfs integration script uses the
configure_networking function, which depends on the IPOPTS environmental
variable set by the init script.  However, this variable is not
available, because it's not exported from init, and local-top/iscsi is
not sourced but executed.  Thus static IP configuration on the kernel
command line is not communicated to the script."

(closes: #516746)

Reported-by: Ferenc Wagner <wferi@niif.hu>