mkinitramfs: add all usb storage devices
USB storage device is not recognized at boot after kernel upgrade
to 2.6.30.
It seems that the usb-storage module has been split into several mini
modules, which are not copied to the initramfs image.

In my case usb-storage.ko is copied but ums-cypress.ko is not copied.

I've attached a patch against hook-functions that fixes this, by using
copy_modules_dir to copy all the modules in kernel/drivers/usb/storage,
instead of just usb-storage.

I've set the severity to important because my USB storage device also
happens to be the boot device...

Cheers,
Avi

(closes: #543568)

Signed-off-by: maximilian attems <maks@debian.org>