mkinitramfs: add LZO support
Albin Tonnerre sent a patch fixing the initramfs LZO decompression,
so heres my fix for initramfs-tools to have compete compression
support. Tell me if you want me to open a new bugreport.

On a side note:	this will only work with a patched kernel (might be
upstream with 2.6.34), its a good idea to include a warning for those
running 2.6.33 from experimental

[ adding it on TODO of Lenny linux-2.6 2.6.32 ]

Signed-off-by: maximilian attems <maks@debian.org>