mkinitramfs: use -9 for lzop
"You do want to use -9 as far as LZO is concerned. The compression ratio is way
better, and the extraction time is actually *not* related to the              
compression level"   Albin Tonnerre

already pointed that out to Maximilian, forgot to put you on CC. Let's
keep the gzip "speed enthusiasts" happy and enable -9 for lzop only.

Signed-off-by: maximilian attems <maks@debian.org>