init: add BOOTIF bootarg
as mentioned by Christoph Bussenius in #535008, it does appear that
setting BOOTIF from the value in /proc/cmdline is needed. at the time
i wrote the patch, i'm not sure it was, but it definitely is now.

with these patchese applied, it seems to work. at least for the moment.

Signed-off-by: maximilian attems <maks@debian.org>