configure_networking(): work with empty DEVICE string
initramfs-tools currently requires a device to be hard-coded, but this
is not much use if the network device is not known ahead of time.  If
the device specified in either /etc/initramfs-tools/initramfs.conf or
on the ip=xxx kernel command line.

usefull for multiple net devices.

Based on patch by Tim Small <tim@seoss.co.uk>

Closes: #566295, #575766

Reviewed-by: Vagrant Cascadian <vagrant@freegeek.org>
Signed-off-by: maximilian attems <maks@debian.org>