hook-functions: make device name handling for /proc/mdstat more flexible for MODULES=dep
We can't assume it's always [hs]dX# in /proc/mdstat.
It can be e.g. also [hs]dX (raw device) or dm-## as well.
So don't statically calculate the string based on
string length but instead try to be as smart as
possible.  Also add support for handling additional
arguments in the /proc/mdstat output like
"(auto-read-only)".

This affects MODULES=dep only.

[Closes: #584520] + [Closes: #514756]

Reported-by: Simon Richter <Simon.Richter@hogyros.de>
Reported-by: Martin Michlmayr <tbm@cyrius.com>
Signed-off-by: Michael Prokop <mika@debian.org>