initramfs-tools: work around 'find -printf %Y' bug
The change to filter out looping or broken symlinks from the initramfs,
written by me for Ubuntu and applied to Debian for #575157, tickles a
bug in find (which I think also had an associated Debian bug, but I've
been unable to find it):

  https://savannah.gnu.org/bugs/index.php?29460

This was fixed in 4.5.8, which is currently only in experimental.  In
any case findutils is very deep in the dependency chain for upgrades and
I believe it would be a good idea to avoid requiring a new version of it
(I've had problems quite recently with strict dependencies on findutils
causing unresolvable dependency loops).  The following patch works
around this bug in a reasonably straightforward way.

[ debian findutils bug was #576772 ]
Signed-off-by: maximilian attems <maks@debian.org>