init: provide validate_init() wrapper to support absolute symlinks.
If /sbin/init is an absolute symlink (e.g. to /bin/systemd) the
readlink check doesn't work as $rootmnt is involved.

The validate_init() shell function takes care of this fact, as
well as when guessing the init binary points to a directory. For
example Ubuntu's upstart has its configuration inside /etc/init
whereas this is considered a valid run_init_process() binary by
the Linux kernel and `test -x' returns true when checking for
executables.

Therefore be more conservative with providing full backwards
compability as well as new added support for absolute symlinks at
the same time.

Closes: #590744