configure_networking(): Look for presence of /tmp/net-*.conf files
when $DEVICE is not set (which is now default), otherwise ipconfig may
receive a valid DHCP response, but fails to break out of the loop and is
called repeatedly.

thanks to Petter Reinholdtsen, i think i've figured out the
initramfs-tools portion of this problem (there may still be outstanding
issues with ipconfig).

tested the attached patch, which seems to address the issue for me at
least.

Closes: #584583

Signed-off-by: maximilian attems <max@stro.at>