mkininitramfs: MODULES=dep fallback to /proc/mounts for rootdev
mount output may not contain proper rootdev on some funky aufs setups:

aufs on / type aufs (rw)

"root is on a usb-flash-disk, in normal operation root is on an aufs,
to prevent to many writes to the flash.
for the dist-upgrade the box was booted without the bootarg
aufs=tmpfs."

Thus if mount parsing shows nothing, try /proc/mounts as fallback.

Closes: #600453
Reported-by: kurt witschi <tmpml@gmx.net>
Tested-by: kurt witschi <tmpml@gmx.net>
Signed-off-by: maximilian attems <max@stro.at>