initramfs-tools: MODULES=dep fix for ubifs /
Similar to #582858, however for me it still does not work
with symlinked /etc/mtab to /proc/mounts.

---- contents of /proc/mounts ----
rootfs / rootfs rw 0 0
ubi0:rootfs / ubifs rw,relatime 0 0

Tighten the regex not to match ^/dev/ but not to match the
rootfs entry while also checking for the `/' root entry.

Also fix the FSTYPE entry to look for the correct field in
/proc/mounts even if not used later.

Closes: #608339
Reported-by: Alexander Clouter <alex@digriz.org.uk>
Tested-by: Alexander Clouter <alex@digriz.org.uk>
Reviewed-by: Michael Prokop <mika@debian.org>
Signed-off-by: maximilian attems <max@stro.at>